feat: Add Toggle Section Behavior
Adds a new setting to allow for highlight toggling, rather than having separate binds for highlighting and unhighlighting. This will also enable text selections to persist, so whatever text you have selected will not be deselected after toggling a highlight.

By default, the setting is disabled, so the default behavior is unchanged.