feat: Add Selectionless Highlighting
Allows for highlighting words just by pressing the keybind while inside the word, similar to obsidian's native text editing tools. If the user has no selections, and triggers a highlight, it will search back and highlight from the beginning to end of the current word the cursor is in.

It will also work for unhighlighting. If the user's cursor is inside of a css highlight snippet, and they trigger an erase (whether through the old command or the new toggle), it will select the highlighted region and apply the erase automatically.